Use pure bytes approach to prevent memory abuse in embedded environments.
During previous debugging, a stack overflow occurred which took considerable time to investigate, eventually traced back to FixedString. We may need to be more cautious when using FixedString - even though it's allocated on the stack when created directly in functions, its accumulation can still lead to memory overflow that's difficult to trace. We may need to limit memory usage within modules, so in this commit I'm attempting to implement CTL200 using a bytes-only approach.

Additionally, this commit includes the following changes:
- Enabled the correct memory.x.
- Due to some issues encountered with DMA1, we're temporarily using BufferedUart to handle uart-related work, to be improved later.
- Removed some previous experimental examples.